refactored some code in PlateRecognizer.cs Form1.cs -> MainForm.cs
