+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,9 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\epicenterWin",
+    "\\packages"
+  ],
+  "SelectedNode": "\\epicenterWin\\MainForm.cs",
+  "PreviewInSolutionExplorer": false
+}
