+++ epicenterWin/Program.cs
@@ -13,7 +13,7 @@ namespace epicenterWin
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new MainForm());
         }
     }
 }
